refactor includes to be framework compliant for Apple
