+++ Core/AppRuntime/Source/AppRuntime.cpp
@@ -1,3 +1,4 @@
+#include <napi/napi.h>
 #include "AppRuntime.h"
 
 #include "WorkQueue.h"
