+++ Core/AppRuntime/Source/AppRuntimeWin32.cpp
@@ -1,3 +1,4 @@
+#include <napi/napi.h>
 #include "AppRuntime.h"
 
 #include <Objbase.h>
