+++ Core/ScriptLoader/Source/ScriptLoader.cpp
@@ -1,3 +1,4 @@
+#include <napi/napi.h>
 #include <Babylon/ScriptLoader.h>
 #include <UrlLib/UrlLib.h>
 #include <arcana/threading/task.h>
