+++ Dependencies/napi/source/env_JavaScriptCore.cc
@@ -1,4 +1,4 @@
-#include <napi/env.h>
+#include <napi/napi.h>
 #include <napi/js_native_api_types.h>
 #include "JavaScriptCore/JavaScript.h"
 #include "js_native_api_JavaScriptCore.h"
