+++ Dependencies/napi/source/js_native_api_JavaScriptCore.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <napi/js_native_api.h>
 #include <napi/js_native_api_types.h>
+#include <napi/js_native_api.h>
 #include <JavaScriptCore/JavaScript.h>
 #include <unordered_set>
 #include <list>
