+++ Plugins/NativeInput/Include/Babylon/Plugins/NativeInput.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <napi/env.h>
-
 namespace Babylon::Plugins
 {
     class NativeInput final
