+++ Plugins/NativeWindow/Include/Babylon/Plugins/NativeWindow.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <napi/env.h>
-
 namespace Babylon::Plugins::NativeWindow
 {
     void Initialize(Napi::Env env, void* windowPtr, size_t width, size_t height);
