+++ Plugins/NativeXr/Include/Babylon/Plugins/NativeXr.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <napi/env.h>
-
 namespace Babylon::Plugins::NativeXr
 {
     void Initialize(Napi::Env env);
