+++ Polyfills/Console/Include/Babylon/Polyfills/Console.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <napi/env.h>
-
 namespace Babylon::Polyfills::Console
 {
     /**
