+++ Polyfills/Window/Include/Babylon/Polyfills/Window.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <napi/env.h>
-
 namespace Babylon::Polyfills::Window
 {
     void Initialize(Napi::Env env);
