+++ Polyfills/XMLHttpRequest/Include/Babylon/Polyfills/XMLHttpRequest.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <napi/env.h>
-
 namespace Babylon::Polyfills::XMLHttpRequest
 {
     void Initialize(Napi::Env env);
